+++ Sheeeeeeeeet/Items/Items/ActionSheetMultiSelectItem.swift
@@ -28,9 +28,19 @@ open class ActionSheetMultiSelectItem: ActionSheetSelectItem {
     
     // MARK: - Initialization
     
-    public override init(title: String, isSelected: Bool, group: String = "", value: Any? = nil, image: UIImage? = nil) {
-        super.init(title: title, isSelected: isSelected, group: group, value: value, image: image)
-        tapBehavior = .none
+    public init(
+        title: String,
+        isSelected: Bool,
+        group: String = "",
+        value: Any? = nil,
+        image: UIImage? = nil) {
+        super.init(
+            title: title,
+            isSelected: isSelected,
+            group: group,
+            value: value,
+            image: image,
+            tapBehavior: .none)
     }