+++ Sheeeeeeeeet/Items/Items/ActionSheetSelectItem.swift
@@ -46,10 +46,20 @@ open class ActionSheetSelectItem: ActionSheetItem {
     
     // MARK: - Initialization
     
-    init(title: String, isSelected: Bool, group: String = "", value: Any? = nil, image: UIImage? = nil) {
+    init(
+        title: String,
+        isSelected: Bool,
+        group: String = "",
+        value: Any? = nil,
+        image: UIImage? = nil,
+        tapBehavior: TapBehavior = .dismiss) {
         self.isSelected = isSelected
         self.group = group
-        super.init(title: title, value: value, image: image)
+        super.init(
+            title: title,
+            value: value,
+            image: image,
+            tapBehavior: tapBehavior)
     }